Rename tripleo_common to tripleo
Created shims in tripleo_common so that existing users of the
library aren't impacted immediately.

Added deprecation warnings to tripleo_common to ask users to move
to tripleo.common.

Tests in triple_common are kept in place to exercise shims.

Moved tripleo_common to tripleo.common and updated it to use new
name.